fix(demux): add additional check for block_num to the inline action query
There is a situation in which two blocks can have the same transaction ID, which I'm having a hard time replicating but I believe occurs when an action throws an error (i.e times out with 3080006). When this situation arises, the MongoActionReader can try to extract the incorrect transaction since the reader is just querying for the transaction ID. This now expands the query to look for both the transaction ID and the block number.